requirements.txt
pyTelegramBotAPI==4.14.0
PyPDF2==3.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas
openpyxl
chardet
google-cloud-vision
pdfplumber